chore: move engine api trait into server
