+++ src/lib.rs
@@ -28,5 +28,5 @@ pub use probe::*;
 mod health;
 pub use health::*;
 
-pub mod engine;
-pub mod payload;
+mod payload;
+pub use payload::*;
