+++ src/server.rs
@@ -3,7 +3,6 @@ use crate::{
     HealthHandle,
     client::rpc::RpcClient,
     debug_api::DebugServer,
-    engine::EngineApiServer,
     payload::{
         NewPayload, NewPayloadV3, NewPayloadV4, OpExecutionPayloadEnvelope, PayloadSource,
         PayloadTraceContext, PayloadVersion,
@@ -18,6 +17,7 @@ use alloy_rpc_types_engine::{
 use alloy_rpc_types_eth::{Block, BlockNumberOrTag};
 use jsonrpsee::RpcModule;
 use jsonrpsee::core::{RegisterMethodError, RpcResult, async_trait};
+use jsonrpsee::proc_macros::rpc;
 use jsonrpsee::types::ErrorObject;
 use jsonrpsee::types::error::INVALID_REQUEST_CODE;
 use metrics::counter;
@@ -229,6 +229,48 @@ impl TryInto<RpcModule<()>> for RollupBoostServer {
     }
 }
 
+#[rpc(server, client)]
+pub trait EngineApi {
+    #[method(name = "engine_forkchoiceUpdatedV3")]
+    async fn fork_choice_updated_v3(
+        &self,
+        fork_choice_state: ForkchoiceState,
+        payload_attributes: Option<OpPayloadAttributes>,
+    ) -> RpcResult<ForkchoiceUpdated>;
+
+    #[method(name = "engine_getPayloadV3")]
+    async fn get_payload_v3(
+        &self,
+        payload_id: PayloadId,
+    ) -> RpcResult<OpExecutionPayloadEnvelopeV3>;
+
+    #[method(name = "engine_newPayloadV3")]
+    async fn new_payload_v3(
+        &self,
+        payload: ExecutionPayloadV3,
+        versioned_hashes: Vec<B256>,
+        parent_beacon_block_root: B256,
+    ) -> RpcResult<PayloadStatus>;
+
+    #[method(name = "engine_getPayloadV4")]
+    async fn get_payload_v4(
+        &self,
+        payload_id: PayloadId,
+    ) -> RpcResult<OpExecutionPayloadEnvelopeV4>;
+
+    #[method(name = "engine_newPayloadV4")]
+    async fn new_payload_v4(
+        &self,
+        payload: OpExecutionPayloadV4,
+        versioned_hashes: Vec<B256>,
+        parent_beacon_block_root: B256,
+        execution_requests: Vec<Bytes>,
+    ) -> RpcResult<PayloadStatus>;
+
+    #[method(name = "eth_getBlockByNumber")]
+    async fn get_block_by_number(&self, number: BlockNumberOrTag, full: bool) -> RpcResult<Block>;
+}
+
 #[async_trait]
 impl EngineApiServer for RollupBoostServer {
     #[instrument(
@@ -435,7 +477,6 @@ impl EngineApiServer for RollupBoostServer {
 #[allow(clippy::complexity)]
 mod tests {
     use super::*;
-    use crate::engine::EngineApiClient;
     use crate::probe::ProbeLayer;
     use crate::proxy::ProxyLayer;
     use alloy_primitives::hex;
